requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
mistralai>=1.0.0
httpx>=0.27.0
pandas>=2.2.0
openpyxl>=3.1.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0.0
reportlab>=4.0
ollama>=0.4.0
pytest>=8.0
